+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## `connpool` UNRELEASED
 
 - Added support for pulling items from the pool with waiting for the local limit.
+- Added a method to attempt pulling an item from the pool without waiting for the global limit.
 
 ## `connpool` 0.1.0 - December 28, 2025